dom/os/test/test_readfile.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>os open test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();

  function runTest(bytes, expected) {
    return new Promise((res, rej) => {
      var workerCode = `
        try {
          var file = navigator.os.open("../../../../dom/os/test/payload.txt",
            navigator.os.RDONLY, navigator.os.IREAD);
          var data = navigator.os.read(file, ` + bytes + `);
          navigator.os.close(file);

          postMessage(JSON.stringify({
            hasFile: !!file,
            data: data.join(',')
          }));
        }
        catch (ex) {
          postMessage("open err: " + ex);
        }
        `;

      var workerBlob = new Blob([workerCode], { type: 'application/javascript' });
      var w = new Worker(URL.createObjectURL(workerBlob));

      w.onmessage = e => {
        is(e.data.indexOf('open err'), -1, 'ran successfully');
        if (e.data.indexOf('open err') !== -1) {
          dump(e.data + '\n');
          return res();
        }

        try {
          var data = JSON.parse(e.data);
        }
        catch(ex) {
          ok(false, fn + ' is valid JSON');
        }

        is(data.hasFile, true, 'hasFile');

        var expectedData = expected.split('').map(c=>c.charCodeAt(0)).join(',');
        is(data.data, expectedData, 'data');

        res();
      };
    });
  }

  SimpleTest.waitForExplicitFinish();

  Promise.all([
    runTest(20, 'This is data that yo'),
    runTest(2000, 'This is data that you can read through the APIs!') // read too much data
  ]).then(() => {
    SimpleTest.finish();
  });
  </script>
</body>
</html>
